test: fix suite
