+++ e2e/tests/content-manager/editview.spec.ts
@@ -10,9 +10,7 @@ test.describe('Edit View', () => {
   });
 
   test.describe('Collection Type', () => {
-    // TODO V5: D&P
-
-    test.skip('A user should be able to navigate to the EditView of the content manager to create, save, publish, unpublish & delete a new entry', async ({
+    test('A user should be able to navigate to the EditView of the content manager to create, save, publish, unpublish & delete a new entry', async ({
       page,
     }) => {
       await page.getByRole('link', { name: 'Content Manager' }).click();
