+++ packages/admin-test-utils/package.json
@@ -52,6 +52,13 @@
       "source": "./src/global-setup.ts",
       "default": "./dist/global-setup.js"
     },
+    "./setup": {
+      "types": "./dist/setup.d.ts",
+      "module": "./dist/setup.mjs",
+      "require": "./dist/setup.js",
+      "source": "./src/setup.ts",
+      "default": "./dist/setup.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "./dist/index.js",
@@ -75,11 +82,13 @@
     "@strapi/pack-up": "workspace:*",
     "@testing-library/jest-dom": "5.16.5",
     "eslint-config-custom": "4.19.0",
+    "jest-environment-jsdom": "29.6.1",
     "tsconfig": "4.19.0"
   },
   "peerDependencies": {
     "@reduxjs/toolkit": "^1.9.7",
-    "@testing-library/jest-dom": "^5.16.5"
+    "@testing-library/jest-dom": "^5.16.5",
+    "jest-environment-jsdom": "^29.0.0"
   },
   "engines": {
     "node": ">=18.0.0 <=20.x.x"
