--- packages/core/admin/admin/src/content-manager/pages/EditView/components/InformationBoxCE.tsx
@@ -1,12 +1,0 @@
-import { Information } from './Information';
-
-const InformationBoxCE = () => {
-  return (
-    <Information.Root>
-      <Information.Title />
-      <Information.Body />
-    </Information.Root>
-  );
-};
-
-export { InformationBoxCE };
